Konta

+++ Frontend1/Controllers/AccountManagementController.cs
@@ -15,10 +15,13 @@ namespace Frontend.Controllers
 	{
 		HttpConnector _httpConnector = HttpConnector.GetInstance();
 		AccountForm _accountForm = new AccountForm();
-		ChangePasswordForm _passwordForm = new ChangePasswordForm();
 		UserEditForm _userForm = new UserEditForm();
 		List<string> _roles;
 
+
+		/// <summary>
+		/// Start kontrolera. Przygotowuje dostępne akcje w formularzu w zależności od roli użytkownika
+		/// </summary>
 		public void Start()
 		{
 			if (AuthManager.GetInstance().UserRoles.Exists((string el) => { return el == "admin"; }))
@@ -27,12 +30,12 @@ namespace Frontend.Controllers
 				string errorMessage = _httpConnector.LastErrorMessage;
 				if (errorMessage == null)
 				{
-					_accountForm.Roles = _roles;
 					_accountForm.DataSource = _httpConnector.GetUsers();
 					errorMessage = _httpConnector.LastErrorMessage;
 					if (errorMessage == null)
 					{
-						_accountForm.EnableUserView = true;
+						_accountForm.SetActions(AddUser, UpdateUser, DeleteUser);
+						_accountForm.ShowDialog();
 					}
 				}
 				if (errorMessage != null)
@@ -40,16 +43,21 @@ namespace Frontend.Controllers
 					MessageBox.Show(errorMessage);
 				}
 			}
-			_accountForm.SetActions(ChangePassword, AddUser, UpdateUser, DeleteUser);
-			_passwordForm.FormClosing += SavePassword;
-			_accountForm.ShowDialog();
+			else
+			{
+				MessageBox.Show("Brak uprawnień");
+			}
 		}
 
+		/// <summary>
+		/// Obsługa zdarzenia dodania użytkonwika. Otwiera formularz dodawania.
+		/// </summary>
+		/// <param name="sender"></param>
+		/// <param name="eventArgs"></param>
 		private void AddUser(object sender, EventArgs eventArgs)
 		{
 			FrontendUserModel user = new FrontendUserModel();
-			bool continueUntilOk = true;
-			while (continueUntilOk)
+			while (true)
 			{
 				_userForm.OpenForAddition(user);
 				if (_userForm.ShowDialog() == DialogResult.OK)
@@ -62,7 +70,6 @@ namespace Frontend.Controllers
 					}
 					else
 					{
-						continueUntilOk = false;
 						_accountForm.DataSource = null;
 						_accountForm.DataSource = _httpConnector.GetUsers();
 						errorMessage = _httpConnector.LastErrorMessage;
@@ -70,22 +77,25 @@ namespace Frontend.Controllers
 						{
 							MessageBox.Show(errorMessage);
 						}
+						break;
 					}
 				}
-				else
-				{
-					continueUntilOk = false;
-				}
+				break;
 			}
 		}
 
+		/// <summary>
+		/// Obsługa zdarzenia edycji użytkownika.
+		/// </summary>
+		/// <param name="sender"></param>
+		/// <param name="eventArgs"></param>
 		private void UpdateUser(object sender, EventArgs eventArgs)
 		{
-			/*FrontendUserModel user = _accountForm.GetSelectedUser() as FrontendUserModel;
+			FrontendUserModel user = _accountForm.GetSelectedUser() as FrontendUserModel;
 			if (user != null)
 			{
-				bool continueUntilOk = true;
-				while (continueUntilOk)
+				user = user.Clone();
+				while(true)
 				{
 					_userForm.OpenForEdit(user);
 					if (_userForm.ShowDialog() == DialogResult.OK)
@@ -98,7 +108,6 @@ namespace Frontend.Controllers
 						}
 						else
 						{
-							continueUntilOk = false;
 							_accountForm.DataSource = null;
 							_accountForm.DataSource = _httpConnector.GetUsers();
 							errorMessage = _httpConnector.LastErrorMessage;
@@ -106,66 +115,51 @@ namespace Frontend.Controllers
 							{
 								MessageBox.Show(errorMessage);
 							}
+							break;
 						}
 					}
 					else
 					{
-						continueUntilOk = false;
+						break;
 					}
 				}
 			}
 			else
 			{
 				MessageBox.Show("Nie wybrano użytkownika");
-			}*/
+			}
 		}
 
 		private void DeleteUser(object sender, EventArgs eventArgs)
 		{
-			/*FrontendUserModel user = _accountForm.GetSelectedUser() as FrontendUserModel;
+			FrontendUserModel user = _accountForm.GetSelectedUser() as FrontendUserModel;
 			if(user != null)
 			{
-				_httpConnector.DeleteUser(user);
-				string errorMessage = _httpConnector.LastErrorMessage;
-				if (errorMessage != null)
+				if(MessageBox.Show("Czy na pewno chcesz usunąć użytkownika " + user.Username + "?", "Potwierdzenie usunięcia", MessageBoxButtons.YesNo) == DialogResult.Yes)
 				{
-					MessageBox.Show(errorMessage);
-				}
-				else
-				{
-					_accountForm.DataSource = null;
-					_accountForm.DataSource = _httpConnector.GetUsers();
-					errorMessage = _httpConnector.LastErrorMessage;
+					_httpConnector.DeleteUser(user);
+					string errorMessage = _httpConnector.LastErrorMessage;
 					if (errorMessage != null)
 					{
 						MessageBox.Show(errorMessage);
+					}
+					else
+					{
+						_accountForm.DataSource = null;
+						_accountForm.DataSource = _httpConnector.GetUsers();
+						errorMessage = _httpConnector.LastErrorMessage;
+						if (errorMessage != null)
+						{
+							MessageBox.Show(errorMessage);
+						}
 					}
 				}
 			}
 			else
 			{
 				MessageBox.Show("Nie wybrano użytkownika");
-			}*/
-		}
-
-		private void ChangePassword(object sender, EventArgs eventArgs)
-		{
-			_passwordForm.ShowDialog();
-		}
-
-		private void SavePassword(object sender, FormClosingEventArgs e)
-		{
-			if(!_passwordForm.CloseWithoutSaving && !e.Cancel)
-			{
-				string password = _passwordForm.GetPassword();
-				_httpConnector.ChangePassword(password);
-				string errorMessage = _httpConnector.LastErrorMessage;
-				if (errorMessage != null)
-				{
-					MessageBox.Show("Zmiana hasła się nie powiodła. Hasło musi mieć co najmniej 6 liter, zawierać jedną cyfrę, jeden znak specjalny i jedną wielką literę.", "Niepowodzenie zmiany hasła");
-					e.Cancel = true;
-				}
 			}
 		}
+
 	}
 }
